\name{Dominance-package}
\alias{Dominance-package}
\alias{Dominance}
\docType{package}
\title{
  Dominance Calculation and Graphs in Animals
}
\description{
The package can calculate ADI (Average Dominance Index) and can build social network graphs with dual directions, can build a Music Notation Graph
}
\details{
The following are sources of information on \pkg{Dominance} package:

  DESCRIPTION file \code{library(help="Dominance")}\cr
   \cr
  This file \code{package?Dominance}\cr
  \cr
  \cr
  Some help files \cr 
  \code{\link{ADI}}\cr
  \code{\link{FDI}}\cr
  \code{\link{Sociogram}}\cr
  \code{\link{Musicnotation}}\cr
  \cr

}

\author{
Knut Krueger , Konstanze Krueger

Maintainer: Who to complain to <Knut.Krueger@equine-science.de>

}


\keyword{ ADI }
\keyword{ Social Network Graph }
\keyword{ Music Notation Graph }
