% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{search.win.lose}
\alias{search.win.lose}
\title{for internal use only}
\usage{
search.win.lose(data_sheet, ...)
}
\description{
for internal use only
}
\author{
Knut krueger

Knut krueger


---- Should be DIRECTLY executable !! ----
-- ==>  Define data, use random,
--	or do  help(data=index)  for the standard data sets.

 The function is currently defined as
'for internal use only'
}
